requirements.txt
pandas
scikit-learn
streamlit
supabase
python-dotenv
openpyxl